requirements.txt
asgiref==3.2.7
bson==0.5.8
dataclasses==0.6
Django==2.2.13
django-rest-framework-mongoengine==3.4.1
djangorestframework==3.11.0
djongo==1.3.2
dnspython==1.16.0
mongoengine==0.20.0
pymongo==3.10.1
python-dateutil==2.8.1
pytz==2020.1
six==1.15.0
sqlparse==0.2.4
